Add build script logic to facilitate the test switchover to [Fact] style
This change is based on Jeremy's POC PR; I have modified the build
scripts to minimize the changes to the individual test build scripts.
In the simplest case the conversion requires just replacing the Main
method with a [Fact]-marked test entrypoint and removing the
OutputType property from the project file.

Empty OutputType property gets automatically replaced with the Library
default in the SDK scripts; this is not a huge deal but I had to
remove the pre-existing hack of OutputType=Library + CLRTestKind=""
implying SharedLibrary. According to my search most Library projects
already have the SharedLibrary CLRTestKind set, for the rest I plan
to set it using a bulk project modification; I think I should be able
to adapt the ILTransform analyzer for this purpose.

Thanks

Tomas
